Move process_key from brains2 main to shared firmware
